Refactor all errors and add further validations to new users' usernames - Usernames cannot start with a symbol, are case insensitive (for   uniqueness) or contain vulgarities

+++ controllers/SuspiciousActivityLogController.js
@@ -1,38 +1,43 @@
 const SuspiciousActivityLog = require("../models/SuspiciousActivityLog");
 const User = require("../models/User");
-
-// Helper function to handle errors
-const handleError = (res, statusCode, message) => {
-  return res.status(statusCode).json({ error: message });
-};
+const { NotFoundError, Error4xx } = require("../common/utils/errorValues");
+const { handleError } = require("../common/utils/errorHandler");
 
 // [GET] Get a suspicious activity log by ID
 exports.getSuspiciousActivityLogById = async (req, res) => {
-  const { logId } = req.params;
+  const logId = req.params.logId;
+
+  if (!logId) {
+    throw new Error4xx("logId parameter is missing from the URL");
+  }
 
   try {
     const log = await SuspiciousActivityLog.findById(logId);
 
     if (!log) {
-      handleError(res, 404, "Suspicious Activity Log not found");
+      throw new NotFoundError("Suspicious Activity Log not found");
     }
 
     res.json(log);
   } catch (error) {
-    handleError(res, 500, error.message ?? "Internal Server Error");
+    handleError(res, error);
   }
 };
 
 // [GET] Get all suspicious activity logs of a single user by UUID
 exports.getAllSuspiciousActivityLogsByUserId = async (req, res) => {
-  const { uuid } = req.params;
+  const uuid = req.params.uuid;
+
+  if (!uuid) {
+    throw new Error4xx("UUID parameter is missing from the URL");
+  }
 
   try {
     // Check if the user with the specified UUID exists
     const user = await User.findOne({ uuid });
 
     if (!user) {
-      handleError(res, 404, "User not found");
+      throw new NotFoundError("Current user not found");
     }
 
     // Retrieve all suspicious activity logs for the user
@@ -40,7 +45,7 @@ exports.getAllSuspiciousActivityLogsByUserId = async (req, res) => {
 
     res.json(logs);
   } catch (error) {
-    handleError(res, 500, error.message ?? "Internal Server Error");
+    handleError(res, error);
   }
 };
 
@@ -52,14 +57,16 @@ exports.createSuspiciousActivityLog = async (req, res) => {
   try {
     // Check if required fields are missing
     if (!uuid || !reason) {
-      handleError(res, 400, "UUID and reason are required fields");
+      throw new Error4xx("UUID and reason are required request body fields");
+    }
+    if (!reportedBy) {
+      throw new Error4xx("reportedBy missing from query parameters");
     }
 
     // Check if the user with the specified UUID exists
     const user = await User.findOne({ uuid });
-
     if (!user) {
-      handleError(res, 404, "User not found");
+      throw new NotFoundError("Current user not found");
     }
 
     const newLogData = {
@@ -67,46 +74,51 @@ exports.createSuspiciousActivityLog = async (req, res) => {
       reportedBy: reportedBy || "System Generated",
       reason,
     };
-
     const newLog = new SuspiciousActivityLog(newLogData);
-
     // Save the new log to the database
     const savedLog = await newLog.save();
-
     res.status(201).json(savedLog);
   } catch (error) {
-    handleError(res, 500, error.message ?? "Internal Server Error");
+    handleError(res, error);
   }
 };
 
 // [DELETE] Delete a suspicious activity log by ID
 exports.deleteSuspiciousActivityLogById = async (req, res) => {
-  const { logId } = req.params;
+  const logId = req.params.logId;
 
   try {
-    const deletedLog = await SuspiciousActivityLog.findByIdAndDelete(logId);
-
-    if (!deletedLog) {
-      handleError(res, 404, "Suspicious Activity Log not found");
+    if (!logId) {
+      throw new Error4xx("logId parameter is missing from the URL");
     }
-
+    const deletedLog = await SuspiciousActivityLog.findByIdAndDelete(logId);
+    if (!deletedLog) {
+      throw new NotFoundError("Suspicious Activity Log not found");
+    }
     res.json({ message: "Suspicious Activity Log deleted successfully" });
   } catch (error) {
-    handleError(res, 500, error.message ?? "Internal Server Error");
+    handleError(res, error);
   }
 };
 
 // [GET] Get suspicious activity logs by date range
 exports.getSuspiciousActivityLogsByDateRange = async (req, res) => {
-  const { uuid } = req.params;
+  const uuid = req.params.uuid;
   const { startDate, endDate } = req.query;
 
   try {
+    if (!uuid) {
+      throw new Error4xx("UUID is a required request body field");
+    }
+    if (!startDate || !endDate) {
+      throw new Error4xx("Start and/or End date missing from query parameters");
+    }
+
     // Check if the user with the specified UUID exists
     const user = await User.findOne({ uuid });
 
     if (!user) {
-      handleError(res, 404, "User not found");
+      throw new NotFoundError("Current user not found");
     }
 
     // Define a date range query
@@ -123,6 +135,6 @@ exports.getSuspiciousActivityLogsByDateRange = async (req, res) => {
 
     res.json(logs);
   } catch (error) {
-    handleError(res, 500, error.message ?? "Internal Server Error");
+    handleError(res, error);
   }
 };
